move util tests to own file
